recommender/pyproject.toml
[tool.poetry]
name = "recommender"
version = "0.1.0"
description = ""
authors = ["khanghy1000 <46063255+khanghy1000@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
gensim = "4.3.3"
LibRecommender = "1.5.1"
numpy = "1.23.4"
pandas = "2.2.3"
scikit-learn = "1.5.2"
scipy = "1.12.0"
tensorflow = "2.10.0"
tensorflow-io-gcs-filesystem = "0.31.0"
torch = "2.4.1"
tqdm = "4.66.5"
sanic = "^24.6.0"
requests = "^2.32.3"
aiohttp = "^3.11.2"
pydantic = "^2.9.2"
ujson = "^5.10.0"
redis = ">=4.2.0"
faiss-cpu = ">=1.5.2"
fastapi = {extras = ["standard"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
